Move panes to higher-level package
