+++ src/components/panes/AssignmentPane.scss
