+++ src/components/panes/InfoPane.jsx
