+++ src/components/panes/InfoPane.scss
