+++ src/components/panes/InputPane.scss
