+++ src/components/panes/PaneBase.scss
