+++ src/components/panes/ReportPane.scss
