+++ src/components/panes/StatsPane.jsx
