+++ src/components/panes/StatsPane.scss
